+++ app/Http/Controllers/TypeEvenementController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\TypeEvenement;
 use Illuminate\Http\Request;
+use App\Models\TypeEvenement;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class TypeEvenementController extends Controller
@@ -12,20 +13,21 @@ class TypeEvenementController extends Controller
     {
         if (organisateurPermission() == true || auth()->user()->usertype == 1) {
             $typeevenements = TypeEvenement::all();
-        return view('admin.page.type.index', compact('typeevenements'));
-        }
-        else
-        {
+            return view('admin.page.type.index', compact('typeevenements'));
+        } else {
             return view('admin.page.index');
         }
-
     }
 
     public function create(Request $request)
     {
         if (organisateurPermission() == true || auth()->user()->usertype == 0) {
+            if (!Auth::check()) {
+                // Utilisateur non authentifié, rediriger ou retourner une réponse appropriée
+                return redirect()->route('login');
+            }
 
-            $validator = Validator::make($request->all(),[
+            $validator = Validator::make($request->all(), [
                 'libelle' => 'required|string',
                 'etat' => 'nullable|in:accepté,rejeté,en attente',
             ]);
@@ -34,48 +36,79 @@ class TypeEvenementController extends Controller
             $typeevenement->libelle = $request->input('libelle');
             $typeevenement->save();
             return view('admin.page.type.create');
-        }else{
+        } else {
             return view('admin.page.index');
         }
-
     }
 
     public function store(Request $request)
     {
-        $request->validate([
-            'libelle' => 'required|string',
-            'etat' => 'nullable'
-        ]);
+        if (organisateurPermission() == true || auth()->user()->usertype == 0) {
+            if (!Auth::check()) {
+                // Utilisateur non authentifié, rediriger ou retourner une réponse appropriée
+                return redirect()->route('login');
+            }
+            $request->validate([
+                'libelle' => 'required|string',
+                'etat' => 'nullable'
+            ]);
 
-        TypeEvenement::create($request->all());
+            TypeEvenement::create($request->all());
 
-        return redirect()->route('type.index');
+            return redirect()->route('type.index');
+        } else {
+            return view('admin.page.index');
+        }
     }
 
     public function edit($id)
     {
-        $typeevenement=TypeEvenement::find($id);
-        return view('admin.page.type.edit', compact('typeevenement'));
+        if (organisateurPermission() == true || auth()->user()->usertype == 0) {
+            if (!Auth::check()) {
+                // Utilisateur non authentifié, rediriger ou retourner une réponse appropriée
+                return redirect()->route('login');
+            }
+            $typeevenement = TypeEvenement::find($id);
+            return view('admin.page.type.edit', compact('typeevenement'));
+        } else {
+            return view('admin.page.index');
+        }
     }
 
-    public function update(Request $request,$id)
+    public function update(Request $request, $id)
     {
-        $typeevenement=TypeEvenement::find($id);
-        $request->validate([
-            'libelle' => 'required|string',
-        ]);
+        if (organisateurPermission() == true || auth()->user()->usertype == 0) {
+            if (!Auth::check()) {
+                // Utilisateur non authentifié, rediriger ou retourner une réponse appropriée
+                return redirect()->route('login');
+            }
+            $typeevenement = TypeEvenement::find($id);
+            $request->validate([
+                'libelle' => 'required|string',
+            ]);
 
-        $typeevenement->update($request->all());
+            $typeevenement->update($request->all());
 
-        return redirect()->route('type.index')->with('success', 'Reclamation created successfully.');
+            return redirect()->route('type.index')->with('success', ' created successfully.');
+        } else {
+            return view('admin.page.index');
+        }
     }
 
     public function destroy($id)
     {
-        $typeevenement=TypeEvenement::find($id);
-        $typeevenement->delete();
+        if (organisateurPermission() == true || auth()->user()->usertype == 0) {
+            if (!Auth::check()) {
+                // Utilisateur non authentifié, rediriger ou retourner une réponse appropriée
+                return redirect()->route('login');
+            }
+            $typeevenement = TypeEvenement::find($id);
+            $typeevenement->delete();
 
-        return redirect()->route('type.index')->with('success', 'Reclamation deleted successfully.');
+            return redirect()->route('type.index')->with('success', ' deleted successfully.');
+        } else {
+            return view('admin.page.index');
+        }
     }
     // public function accepter($id)
     // {
